Update to work with the latest versions of buildbot.
We were broken with the latest versions of buildbot, though not
horribly. It's been a while since we last updated.

Two things were imported that no longer exist. Fortunately, we didn't
use either of them.

UploadFile no longer supports a WithProperties as the file path
parameter, but we didn't really need it apparently.