Allow schedulers to be excluded from trigger lists.
A BuildTarget can trigger another target, but we may only want this to
happen to certain target names. We currently use a generic name, and
build the scheduler to trigger around that name, in order to have a wide
variety of versions to test with. However, in some cases, we don't want
to bother triggering a given scheduler name. We can now specify triggers
to exclude using trigger_excludes= on BuildTarget.